@@ -9,12 +9,16 @@
     "start": "next start"
   },
   "dependencies": {
-    "next": "14.2.5",
+    "next": "^15.0.0",
     "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "react-dom": "18.3.1",
+    "mapbox-gl": "^3.6.0"
   },
   "devDependencies": {
     "typescript": "^5.4.0",
-    "@cloudflare/next-on-pages": "^1.11.5"
+    "@cloudflare/next-on-pages": "^1.13.16",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@types/node": "^20.14.2"
   }
 }
