@@ -1,6 +1,10 @@
 'use client'
 
 import { useEffect, useMemo, useRef, useState } from 'react'
+import mapboxgl from 'mapbox-gl'
+import 'mapbox-gl/dist/mapbox-gl.css'
+
+export const runtime = 'edge'
 
 type AlertState = {
   status: 'active' | 'ended' | 'timeout'
@@ -9,12 +13,20 @@ type AlertState = {
   permissions: { can_call: boolean; can_reply: boolean; can_call_police: boolean }
 }
 
-export default function ReceiverPage({ params }: { params: { token: string } }) {
-  const { token } = params
+export default function ReceiverPage({ params }: any) {
+  const { token } = (params || {}) as { token: string }
   const [state, setState] = useState<AlertState | null>(null)
   const [error, setError] = useState<string | null>(null)
   const apiBase = process.env.NEXT_PUBLIC_API_BASE || ''
   const esRef = useRef<EventSource | null>(null)
+  const mapRef = useRef<HTMLDivElement | null>(null)
+  const mapInstance = useRef<mapboxgl.Map | null>(null)
+  const markerRef = useRef<mapboxgl.Marker | null>(null)
+  const circleSourceId = 'accuracy-circle'
+  const routeSourceId = 'route-line'
+  const routeCoordsRef = useRef<[number, number][]>([])
+  const mapboxToken = process.env.NEXT_PUBLIC_MAPBOX_TOKEN
+  const lastEventAtRef = useRef<number>(0)
 
   useEffect(() => {
     let closed = false
@@ -24,6 +36,16 @@ export default function ReceiverPage({ params }: { params: { token: string } }) 
         if (!res.ok) throw new Error('failed to load')
         const data = (await res.json()) as AlertState
         if (!closed) setState(data)
+        // Load initial history (route)
+        try {
+          const r = await fetch(`${apiBase}/public/alert/${encodeURIComponent(token)}/locations?limit=200&order=asc`)
+          if (r.ok) {
+            const j = (await r.json()) as { items: { lat: number; lng: number }[] }
+            routeCoordsRef.current = j.items.map((x) => [x.lng, x.lat])
+            // If map is ready, reflect immediately
+            if (mapInstance.current) updateRoute(mapInstance.current)
+          }
+        } catch {}
       } catch (e) {
         if (!closed) setError('読み込みに失敗しました')
       }
@@ -35,22 +57,240 @@ export default function ReceiverPage({ params }: { params: { token: string } }) 
   }, [apiBase, token])
 
   useEffect(() => {
-    const es = new EventSource(`${apiBase}/public/alert/${encodeURIComponent(token)}/stream`)
-    esRef.current = es
-    es.onmessage = (ev) => {
+    let stopped = false
+    let retryMs = 1000
+    function connect() {
+      if (stopped) return
+      const es = new EventSource(`${apiBase}/public/alert/${encodeURIComponent(token)}/stream`)
+      esRef.current = es
+      es.onmessage = (ev) => {
+        try {
+          const evt = JSON.parse(ev.data)
+          lastEventAtRef.current = Date.now()
+          if (evt.type === 'location') setState((s) => (s ? { ...s, latest: evt.latest } : s))
+          if (evt.type === 'status') setState((s) => (s ? { ...s, status: evt.status } : s))
+          // reset backoff on successful message
+          retryMs = 1000
+        } catch {}
+      }
+      es.onerror = () => {
+        es.close()
+        if (stopped) return
+        const wait = retryMs
+        retryMs = Math.min(retryMs * 2, 30000)
+        setTimeout(connect, wait)
+      }
+    }
+    connect()
+    return () => {
+      stopped = true
+      esRef.current?.close()
+    }
+  }, [apiBase, token])
+
+  // Polling fallback: SSEが8秒以上沈黙している時だけ5秒間隔で最新を取得
+  useEffect(() => {
+    let active = true
+    const id = setInterval(async () => {
+      if (!active) return
+      const silentFor = Date.now() - (lastEventAtRef.current || 0)
+      if (lastEventAtRef.current !== 0 && silentFor < 8000) return
       try {
-        const evt = JSON.parse(ev.data)
-        if (evt.type === 'location') setState((s) => (s ? { ...s, latest: evt.latest } : s))
-        if (evt.type === 'status') setState((s) => (s ? { ...s, status: evt.status } : s))
+        const res = await fetch(`${apiBase}/public/alert/${encodeURIComponent(token)}`)
+        if (!res.ok) return
+        const data = (await res.json()) as AlertState
+        setState((prev) => {
+          const prevTs = prev?.latest ? new Date(prev.latest.captured_at).getTime() : 0
+          const nextTs = data.latest ? new Date(data.latest.captured_at).getTime() : 0
+          if (nextTs > prevTs && data.latest) {
+            appendRoute(data.latest.lng, data.latest.lat)
+            if (mapInstance.current) updateRoute(mapInstance.current)
+            return prev ? { ...prev, latest: data.latest } : data
+          }
+          return prev ?? data
+        })
       } catch {}
+    }, 5000)
+    return () => {
+      active = false
+      clearInterval(id)
     }
-    es.onerror = () => {
-      es.close()
-    }
-    return () => es.close()
   }, [apiBase, token])
 
   const remaining = useMemo(() => (state ? Math.max(0, state.remaining_sec) : 0), [state])
+
+  // Initialize Mapbox map when token and container are ready
+  useEffect(() => {
+    if (!mapRef.current) return
+    if (!mapboxToken) return
+    if (mapInstance.current) return
+    mapboxgl.accessToken = mapboxToken
+    const map = new mapboxgl.Map({
+      container: mapRef.current,
+      style: 'mapbox://styles/mapbox/streets-v12',
+      center: [139.767, 35.681],
+      zoom: 14,
+    })
+    map.addControl(new mapboxgl.NavigationControl({ showCompass: false }), 'top-right')
+    map.on('load', () => {
+      // Localize labels to Japanese where available
+      try { localizeLabelsToJapanese(map) } catch {}
+      // Prepare accuracy circle source/layer
+      if (!map.getSource(circleSourceId)) {
+        map.addSource(circleSourceId, { type: 'geojson', data: emptyCircle() })
+        map.addLayer({
+          id: 'accuracy-fill',
+          type: 'fill',
+          source: circleSourceId,
+          paint: { 'fill-color': '#3b82f6', 'fill-opacity': 0.15 },
+        })
+        map.addLayer({
+          id: 'accuracy-outline',
+          type: 'line',
+          source: circleSourceId,
+          paint: { 'line-color': '#3b82f6', 'line-width': 1 },
+        })
+      }
+      // Prepare route line source/layer
+      if (!map.getSource(routeSourceId)) {
+        map.addSource(routeSourceId, { type: 'geojson', data: emptyRoute() })
+        map.addLayer({
+          id: 'route-line',
+          type: 'line',
+          source: routeSourceId,
+          paint: { 'line-color': '#2563eb', 'line-width': 3 },
+          layout: { 'line-cap': 'round', 'line-join': 'round' },
+        })
+      }
+      // Place marker if we have initial location
+      const latest = state?.latest
+      if (latest) {
+        upsertMarker(latest.lng, latest.lat)
+        updateCircle(map, latest.lat, latest.lng, latest.accuracy_m ?? 0)
+        appendRoute(latest.lng, latest.lat)
+        updateRoute(map)
+        map.jumpTo({ center: [latest.lng, latest.lat], zoom: 15 })
+      }
+    })
+    mapInstance.current = map
+    return () => {
+      map.remove()
+      mapInstance.current = null
+      markerRef.current = null
+    }
+  }, [mapboxToken, state?.latest])
+
+  // Update map when latest location changes
+  useEffect(() => {
+    const latest = state?.latest
+    const map = mapInstance.current
+    if (!latest || !map) return
+    upsertMarker(latest.lng, latest.lat)
+    updateCircle(map, latest.lat, latest.lng, latest.accuracy_m ?? 0)
+    appendRoute(latest.lng, latest.lat)
+    updateRoute(map)
+    // Smoothly move for subsequent updates
+    map.easeTo({ center: [latest.lng, latest.lat], duration: 800 })
+  }, [state?.latest])
+
+  function upsertMarker(lng: number, lat: number) {
+    if (!markerRef.current) {
+      markerRef.current = new mapboxgl.Marker({ color: '#ef4444' }).setLngLat([lng, lat]).addTo(mapInstance.current!)
+    } else {
+      markerRef.current.setLngLat([lng, lat])
+    }
+  }
+
+  function updateCircle(map: mapboxgl.Map, lat: number, lng: number, accuracy_m: number) {
+    const source = map.getSource(circleSourceId) as mapboxgl.GeoJSONSource | undefined
+    if (!source) return
+    const radius = Math.max(accuracy_m || 0, 0)
+    const data = radius > 0 ? circleGeoJSON(lat, lng, radius) : emptyCircle()
+    source.setData(data)
+  }
+
+  function emptyCircle(): GeoJSON.FeatureCollection<GeoJSON.Polygon> {
+    return { type: 'FeatureCollection', features: [] }
+  }
+
+  function circleGeoJSON(lat: number, lng: number, radiusMeters: number): GeoJSON.FeatureCollection<GeoJSON.Polygon> {
+    const points = 64
+    const coords: [number, number][] = []
+    const R = 6378137
+    for (let i = 0; i <= points; i++) {
+      const theta = (i / points) * 2 * Math.PI
+      const dx = (radiusMeters * Math.cos(theta)) / (R * Math.cos((lat * Math.PI) / 180))
+      const dy = radiusMeters * Math.sin(theta) / R
+      const lngOffset = (dx * 180) / Math.PI
+      const latOffset = (dy * 180) / Math.PI
+      coords.push([lng + lngOffset, lat + latOffset])
+    }
+    return {
+      type: 'FeatureCollection',
+      features: [
+        {
+          type: 'Feature',
+          geometry: { type: 'Polygon', coordinates: [coords] },
+          properties: {},
+        },
+      ],
+    }
+  }
+
+  function emptyRoute(): GeoJSON.FeatureCollection<GeoJSON.LineString> {
+    return { type: 'FeatureCollection', features: [] }
+  }
+
+  function appendRoute(lng: number, lat: number) {
+    const coords = routeCoordsRef.current
+    const last = coords[coords.length - 1]
+    // Avoid pushing duplicate points
+    if (!last || last[0] !== lng || last[1] !== lat) {
+      coords.push([lng, lat])
+      // Keep last 200 points to bound memory
+      if (coords.length > 200) coords.shift()
+    }
+  }
+
+  function updateRoute(map: mapboxgl.Map) {
+    const coords = routeCoordsRef.current
+    const src = map.getSource(routeSourceId) as mapboxgl.GeoJSONSource | undefined
+    if (!src) return
+    if (coords.length < 2) {
+      src.setData(emptyRoute())
+      return
+    }
+    const data: GeoJSON.FeatureCollection<GeoJSON.LineString> = {
+      type: 'FeatureCollection',
+      features: [
+        { type: 'Feature', geometry: { type: 'LineString', coordinates: coords }, properties: {} },
+      ],
+    }
+    src.setData(data)
+  }
+
+  // Replace text labels with Japanese names where available
+  function localizeLabelsToJapanese(map: mapboxgl.Map) {
+    const style = map.getStyle()
+    if (!style?.layers) return
+    for (const layer of style.layers) {
+      if (layer.type !== 'symbol') continue
+      const id = layer.id
+      // text-field があるレイヤのみ対象
+      // 既存の複雑なformat式のレイヤには適用しない（盾アイコン等を避ける）
+      const tf = (layer as any).layout?.['text-field']
+      if (!tf) continue
+      // シールド系やハイウェイ番号などはスキップ
+      if (id.includes('shield') || id.includes('motorway') && String(tf).includes('{reflen}')) continue
+      try {
+        map.setLayoutProperty(id, 'text-field', [
+          'coalesce',
+          ['get', 'name_ja'],
+          ['get', 'name']
+        ])
+      } catch {}
+    }
+  }
 
   return (
     <main style={{ padding: 16, display: 'grid', gap: 12 }}>
@@ -65,16 +305,13 @@ export default function ReceiverPage({ params }: { params: { token: string } }) 
             最終更新: {state.latest ? new Date(state.latest.captured_at).toLocaleString() : '—'} / バッテリー:{' '}
             {state.latest?.battery_pct ?? '—'}%
           </div>
-          <div style={{ height: 280, background: '#f3f4f6', borderRadius: 8, display: 'grid', placeItems: 'center' }}>
-            <div>
-              地図プレースホルダ
-              {state.latest && (
-                <div style={{ fontSize: 12, opacity: 0.7 }}>
-                  lat {state.latest.lat.toFixed(5)} / lng {state.latest.lng.toFixed(5)} ±
-                  {state.latest.accuracy_m ?? '—'}m
-                </div>
-              )}
-            </div>
+          <div style={{ height: 320, borderRadius: 8, overflow: 'hidden', position: 'relative', background: '#e5e7eb' }}>
+            {!mapboxToken && (
+              <div style={{ position: 'absolute', inset: 0, display: 'grid', placeItems: 'center', zIndex: 1, background: 'rgba(255,255,255,0.8)' }}>
+                <div style={{ color: '#111827' }}>地図トークンが未設定です（NEXT_PUBLIC_MAPBOX_TOKEN）</div>
+              </div>
+            )}
+            <div ref={mapRef} style={{ width: '100%', height: '100%' }} />
           </div>
           <div style={{ display: 'flex', gap: 8 }}>
             <a href="tel:0000000000" style={btn()}>電話</a>
